Removed extra test file
